Made scale limited to int

+++ Minigin/Renderer.cpp
@@ -56,7 +56,7 @@ void dae::Renderer::RenderTexture(const Texture2D& texture, const float x, const
 	SDL_RenderCopy(GetSDLRenderer(), texture.GetSDLTexture(), nullptr, &dst);
 }
 
-void dae::Renderer::RenderTexture(const Texture2D& texture, const float x, const float y, const float scale) const
+void dae::Renderer::RenderTexture(const Texture2D& texture, const float x, const float y, const int scale) const
 {
 	SDL_Rect dst{};
 	dst.x = static_cast<int>(x);
